+++ bibapp/Connector/Requests/BAURLRequestService.m
@@ -62,7 +62,7 @@ static id<BAURLRequestService> sharedInstance;
    [theRequest setHTTPMethod:@"POST"];
    [theRequest setValue:@"application/json" forHTTPHeaderField:@"Accept"];
    [theRequest setValue:@"application/json" forHTTPHeaderField:@"Content-Type"];
-   [theRequest setValue:[NSString stringWithFormat:@"%d", contentLength] forHTTPHeaderField:@"Content-Length"];
+   [theRequest setValue:[NSString stringWithFormat:@"%lu", (unsigned long)contentLength] forHTTPHeaderField:@"Content-Length"];
    [theRequest setHTTPBody:body];
    
    return theRequest;
